Chandler's Code from Fall 2014
The focus of this code was to obtain the sensitivity of the experiment.
It successfully can fit data created from PDFs of the single electron
energy, summed energy, and angular distributions.  The "Test_Stat" code
is for the profile likelihood ratio test statistic.  "Test_Stat_Dist"
creates a distribution of the test stats.  In attempting to be more
organized, I switched to creating workspaces/models, and used them in
SimpleHypoTest and StandardHypotTestInvDemo which are tutorials.  These
should be doing the same thing as my "Test_Stat" code, but I'm not
convinced yet, and haven't fully figured out how the functions the
tutorials are using work.